Struct try_runtime_cli::TryRuntimeCmd

pub struct TryRuntimeCmd {
    pub shared: SharedParams,
    pub command: Command,
}

Our try-runtime command.

See Command for more info.

Fields
shared: SharedParamscommand: Command
Implementations

impl TryRuntimeCmd

pub async fn run<Block, HostFns, BBIP>( &self, block_building_info_provider: Option<BBIP> ) -> Result<()>where Block: BlockT<Hash = H256> + DeserializeOwned, Block::Header: DeserializeOwned, <Block::Hash as FromStr>::Err: Debug, <NumberFor<Block> as FromStr>::Err: Debug, <NumberFor<Block> as TryInto<u64>>::Error: Debug, NumberFor<Block>: FromStr, HostFns: HostFunctions, BBIP: BlockBuildingInfoProvider<Block, Option<(InherentData, Digest)>>,

๐Deprecated: Substrateโs try-runtime subcommand has been migrated to a standalone CLI (https://github.com/paritytech/try-runtime-cli). It is no longer being maintained here and will be removed entirely some time after January 2024. Please remove this subcommand from your runtime and use the standalone CLI.