Crate try_runtime_cli

Try-runtime

Substrate’s try-runtime subcommand has been migrated to a standalone CLI.

It is no longer maintained here and will be removed in the future.

Modules

block_building_info

commands

Structs

LiveState

A Live variant State

SharedParams

Shared parameters of the try-runtime commands

TryRuntimeCmd

Our try-runtime command.

Enums

Command

Possible commands of try-runtime.

Runtime

State

The source of runtime state to use.

Constants

DEPRECATION_NOTICE